refactor for maintance - splitting up modules, score method to evalue, gpu settings consildated
