+++ denseclus/__init__.py
@@ -1,14 +1,10 @@
 """
 Dense Clus Library
-
-Authors: Charles Frenzel, Baichaun Sun
-Date: November 2023
 """
 
+from .categorical import extract_categorical
 from .DenseClus import DenseClus
-from .utils import extract_categorical, extract_numerical
-
-__version__ = "0.2.0"
+from .numerical import extract_numerical
 
 if __name__ == "__main__":  # pragma: no cover
     print(type(DenseClus), type(extract_categorical), type(extract_numerical))
