+++ denseclus/__version__.py
@@ -1,0 +1,7 @@
+# pylint: skip-file
+__title__ = "amazon_denseclus"
+__description__ = "Dense Clustering for Mixed Data Types"
+__version__ = "0.2.1"
+__url__ = ("https://github.com/awslabs/amazon-denseclus",)
+__author__ = "Charles Frenzel & Baichuan Sun"
+__license__ = "MIT"
